backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-genai==1.0.0
websockets==12.0
python-multipart==0.0.9
pydantic==2.9.2
